Move graph test case into /mock folder
