+++ x-pack/solutions/security/packages/kbn-cloud-security-posture/graph/src/components/minimap/minimap_props.stories.tsx
@@ -30,7 +30,7 @@ import {
   RectangleNode,
 } from '../node';
 import { DefaultEdge } from '../edge';
-import { graphSample } from './graph_test_case';
+import { graphSample } from '../mock/graph_sample';
 
 const nodeTypes = {
   hexagon: HexagonNode,
