+++ x-pack/solutions/security/packages/kbn-cloud-security-posture/graph/src/components/mock/graph_sample.ts
